requirements.txt
pytest
flask
python-dotenv
boto3
python-multipart
flask_cors
psycopg2-binary
Flask-Limiter
redis>=4.5.4
marshmallow
sqlalchemy
alembic
gunicorn
bcrypt

# Dev tools
black
isort
flake8
mypy
pre-commit
